+++ build_scripts/qt/sources.pri
@@ -14,7 +14,7 @@ SOURCES += src/main.cpp\
     src/tabcontrollers/audiomanager/AudioManagerDummy.cpp \
     src/tabcontrollers/keyboardinput/KeyboardInputDummy.cpp \
     src/openvr/openvr_init.cpp \
-    src/ivrinput/ivrinput.cpp
+    src/openvr/ivrinput.cpp
 
 HEADERS += src/overlaycontroller.h \
     src/tabcontrollers/AudioTabController.h \
@@ -35,12 +35,11 @@ HEADERS += src/overlaycontroller.h \
     src/tabcontrollers/audiomanager/AudioManagerDummy.h \
     src/tabcontrollers/keyboardinput/KeyboardInputDummy.h \
     src/openvr/openvr_init.h \
-    src/ivrinput/ivrinput_action.h \
-    src/ivrinput/ivrinput_manifest.h \
-    src/ivrinput/ivrinput_action_set.h \
-    src/ivrinput/ivrinput.h \
+    src/openvr/ivrinput_action.h \
+    src/openvr/ivrinput_manifest.h \
+    src/openvr/ivrinput_action_set.h \
+    src/openvr/ivrinput.h \
     src/utils/setup.h
-
 
 win32 {
     SOURCES += src/tabcontrollers/audiomanager/AudioManagerWindows.cpp \
